Move image path models to common domain package
