+++ face/capture/src/main/java/com/simprints/face/capture/FaceCaptureResult.kt
@@ -1,8 +1,8 @@
 package com.simprints.face.capture
 
 import androidx.annotation.Keep
+import com.simprints.core.domain.image.SecuredImageRef
 import com.simprints.core.domain.step.StepResult
-import com.simprints.infra.images.model.SecuredImageRef
 
 @Keep
 data class FaceCaptureResult(
