+++ face/capture/src/main/java/com/simprints/face/capture/models/FaceDetection.kt
@@ -1,9 +1,9 @@
 package com.simprints.face.capture.models
 
 import android.graphics.Bitmap
+import com.simprints.core.domain.image.SecuredImageRef
 import com.simprints.core.tools.time.Timestamp
 import com.simprints.face.infra.basebiosdk.detection.Face
-import com.simprints.infra.images.model.SecuredImageRef
 import java.util.UUID
 
 internal data class FaceDetection(
