+++ face/capture/src/test/java/com/simprints/face/capture/usecases/SaveFaceSampleUseCaseTest.kt
@@ -1,10 +1,10 @@
 package com.simprints.face.capture.usecases
 
 import com.google.common.truth.Truth.assertThat
+import com.simprints.core.domain.image.Path
+import com.simprints.core.domain.image.SecuredImageRef
 import com.simprints.infra.events.session.SessionEventRepository
 import com.simprints.infra.images.ImageRepository
-import com.simprints.infra.images.model.Path
-import com.simprints.infra.images.model.SecuredImageRef
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
 import io.mockk.coVerify
