+++ infra/core/src/main/java/com/simprints/core/domain/image/SecuredImageRef.kt
@@ -1,0 +1,9 @@
+package com.simprints.core.domain.image
+
+import androidx.annotation.Keep
+import java.io.Serializable
+
+@Keep
+data class SecuredImageRef(
+    val relativePath: Path,
+) : Serializable
