+++ infra/images/src/main/java/com/simprints/infra/images/local/ImageLocalDataSourceImpl.kt
@@ -2,8 +2,8 @@ package com.simprints.infra.images.local
 
 import android.content.Context
 import com.simprints.core.DispatcherIO
-import com.simprints.infra.images.model.Path
-import com.simprints.infra.images.model.SecuredImageRef
+import com.simprints.core.domain.image.Path
+import com.simprints.core.domain.image.SecuredImageRef
 import com.simprints.infra.logging.Simber
 import com.simprints.infra.security.SecurityManager
 import dagger.hilt.android.qualifiers.ApplicationContext
