--- infra/images/src/main/java/com/simprints/infra/images/model/ImageRef.kt
@@ -1,7 +1,0 @@
-package com.simprints.infra.images.model
-
-import java.io.Serializable
-
-abstract class ImageRef(
-    open val relativePath: Path,
-) : Serializable
