--- infra/images/src/main/java/com/simprints/infra/images/model/SecuredImageRef.kt
@@ -1,8 +1,0 @@
-package com.simprints.infra.images.model
-
-import androidx.annotation.Keep
-
-@Keep
-data class SecuredImageRef(
-    override val relativePath: Path,
-) : ImageRef(relativePath)
