+++ infra/images/src/test/java/com/simprints/infra/images/remote/signedurl/usecase/FetchUploadUrlsPerSampleUseCaseTest.kt
@@ -1,9 +1,9 @@
 package com.simprints.infra.images.remote.signedurl.usecase
 
 import com.google.common.truth.Truth.*
+import com.simprints.core.domain.image.Path
+import com.simprints.core.domain.image.SecuredImageRef
 import com.simprints.infra.authstore.AuthStore
-import com.simprints.infra.images.model.Path
-import com.simprints.infra.images.model.SecuredImageRef
 import com.simprints.infra.images.remote.signedurl.SampleUploadData
 import com.simprints.infra.images.remote.signedurl.api.ApiSampleUploadUrlRequest
 import com.simprints.infra.images.remote.signedurl.api.ApiSampleUploadUrlResponse
